requirements.txt
asgiref==3.11.0
Django==6.0
gunicorn==23.0.0
Pillow==10.4.0
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.7.0